+++ include/aws/auth/signing_config.h
@@ -62,11 +62,11 @@ enum aws_signature_type {
 
     /* *********************************** come back to this when better understand requirments ***********************/
     /**
-     * Compute a signature for the payload trailing headers.  The signable's input stream should be the chunk data and
+     * Compute a signature for the payload chunked trailer.  The signable's input stream should be the chunk data and
      * the signable should contain the most recent signature value (either the original http request or the most recent
      * chunk) in the "previous-signature" property.
      */
-    AWS_ST_HTTP_REQUEST_TRAILING_HEADERS,
+    AWS_ST_HTTP_REQUEST_CHUNKED_TRAILER,
 
     /**
      * Compute a signature for an event stream event.  The signable's input stream should be the event payload, the
