+++ source/signable_trailer.c
@@ -1,0 +1,112 @@
+/**
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
+#include <aws/auth/signable.h>
+#include <aws/common/string.h>
+
+struct aws_signable_trailing_headers_impl { /* <------------------ reference here -------------------- */
+    struct aws_input_stream *trailing_headers_data;
+    struct aws_string *previous_signature;
+};
+
+static int s_aws_signable_trailing_headers_get_property(
+    const struct aws_signable *signable,
+    const struct aws_string *name,
+    struct aws_byte_cursor *out_value) {
+
+    struct aws_signable_trailing_headers_impl *impl = signable->impl;
+
+    AWS_ZERO_STRUCT(*out_value);
+
+    /*
+     * uri and method can be queried directly from the wrapper request
+     */
+    if (aws_string_eq(name, g_aws_previous_signature_property_name)) {
+        *out_value = aws_byte_cursor_from_string(impl->previous_signature);
+    } else {
+        return AWS_OP_ERR;
+    }
+
+    return AWS_OP_SUCCESS;
+}
+
+static int s_aws_signable_trailing_headers_get_property_list(
+    const struct aws_signable *signable,
+    const struct aws_string *name,
+    struct aws_array_list **out_list) {
+    (void)signable;
+    (void)name;
+    (void)out_list;
+
+    return AWS_OP_ERR;
+}
+
+static int s_aws_signable_trailing_headers_get_payload_stream(
+    const struct aws_signable *signable,
+    struct aws_input_stream **out_input_stream) {
+
+    struct aws_signable_trailing_headers_impl *impl = signable->impl;
+    *out_input_stream = impl->trailing_headers_data;
+
+    return AWS_OP_SUCCESS;
+}
+
+static void s_aws_signable_trailing_headers_destroy(struct aws_signable *signable) {
+    if (signable == NULL) {
+        return;
+    }
+
+    struct aws_signable_trailing_headers_impl *impl = signable->impl;
+    if (impl == NULL) {
+        return;
+    }
+
+    aws_string_destroy(impl->previous_signature);
+
+    aws_mem_release(signable->allocator, signable);
+}
+
+static struct aws_signable_vtable s_signable_trailing_headers_vtable = {
+    .get_property = s_aws_signable_trailing_headers_get_property,
+    .get_property_list = s_aws_signable_trailing_headers_get_property_list,
+    .get_payload_stream = s_aws_signable_trailing_headers_get_payload_stream,
+    .destroy = s_aws_signable_trailing_headers_destroy,
+};
+
+struct aws_signable *aws_signable_new_trailing_headers(
+    struct aws_allocator *allocator,
+    struct aws_input_stream *trailing_headers_data,
+    struct aws_byte_cursor previous_signature) {
+
+    struct aws_signable *signable = NULL;
+    struct aws_signable_trailing_headers_impl *impl = NULL;
+    aws_mem_acquire_many(
+        allocator, 2, &signable, sizeof(struct aws_signable), &impl, sizeof(struct aws_signable_trailing_headers_impl));
+
+    if (signable == NULL || impl == NULL) {
+        return NULL;
+    }
+
+    AWS_ZERO_STRUCT(*signable);
+    AWS_ZERO_STRUCT(*impl);
+
+    signable->allocator = allocator;
+    signable->vtable = &s_signable_trailing_headers_vtable;
+    signable->impl = impl;
+
+    impl->trailing_headers_data = trailing_headers_data;
+    impl->previous_signature = aws_string_new_from_array(allocator, previous_signature.ptr, previous_signature.len);
+    if (impl->previous_signature == NULL) {
+        goto on_error;
+    }
+
+    return signable;
+
+on_error:
+
+    aws_signable_destroy(signable);
+
+    return NULL;
+}
