+++ source/sigv4_http_request.c
@@ -166,237 +166,3 @@ int aws_apply_signing_result_to_http_request(
 
     return AWS_OP_SUCCESS;
 }
-
-/*
- * This is a simple aws_signable wrapper implementation for the aws_http_message struct
- */
-struct aws_signable_http_request_impl {
-    struct aws_http_message *request;
-    struct aws_array_list headers;
-};
-
-static int s_aws_signable_http_request_get_property(
-    const struct aws_signable *signable,
-    const struct aws_string *name,
-    struct aws_byte_cursor *out_value) {
-
-    struct aws_signable_http_request_impl *impl = signable->impl;
-
-    AWS_ZERO_STRUCT(*out_value);
-
-    /*
-     * uri and method can be queried directly from the wrapper request
-     */
-    if (aws_string_eq(name, g_aws_http_uri_property_name)) {
-        aws_http_message_get_request_path(impl->request, out_value);
-    } else if (aws_string_eq(name, g_aws_http_method_property_name)) {
-        aws_http_message_get_request_method(impl->request, out_value);
-    } else {
-        return AWS_OP_ERR;
-    }
-
-    return AWS_OP_SUCCESS;
-}
-
-static int s_aws_signable_http_request_get_property_list(
-    const struct aws_signable *signable,
-    const struct aws_string *name,
-    struct aws_array_list **out_list) {
-
-    struct aws_signable_http_request_impl *impl = signable->impl;
-
-    *out_list = NULL;
-
-    if (aws_string_eq(name, g_aws_http_headers_property_list_name)) {
-        *out_list = &impl->headers;
-    } else {
-        return AWS_OP_ERR;
-    }
-
-    return AWS_OP_SUCCESS;
-}
-
-static int s_aws_signable_http_request_get_payload_stream(
-    const struct aws_signable *signable,
-    struct aws_input_stream **out_input_stream) {
-
-    struct aws_signable_http_request_impl *impl = signable->impl;
-    *out_input_stream = aws_http_message_get_body_stream(impl->request);
-
-    return AWS_OP_SUCCESS;
-}
-
-static void s_aws_signable_http_request_destroy(struct aws_signable *signable) {
-    if (signable == NULL) {
-        return;
-    }
-
-    struct aws_signable_http_request_impl *impl = signable->impl;
-    if (impl == NULL) {
-        return;
-    }
-
-    aws_array_list_clean_up(&impl->headers);
-    aws_mem_release(signable->allocator, signable);
-}
-
-static struct aws_signable_vtable s_signable_http_request_vtable = {
-    .get_property = s_aws_signable_http_request_get_property,
-    .get_property_list = s_aws_signable_http_request_get_property_list,
-    .get_payload_stream = s_aws_signable_http_request_get_payload_stream,
-    .destroy = s_aws_signable_http_request_destroy,
-};
-
-struct aws_signable *aws_signable_new_http_request(struct aws_allocator *allocator, struct aws_http_message *request) {
-
-    struct aws_signable *signable = NULL;
-    struct aws_signable_http_request_impl *impl = NULL;
-    aws_mem_acquire_many(
-        allocator, 2, &signable, sizeof(struct aws_signable), &impl, sizeof(struct aws_signable_http_request_impl));
-
-    if (signable == NULL || impl == NULL) {
-        return NULL;
-    }
-
-    AWS_ZERO_STRUCT(*signable);
-    AWS_ZERO_STRUCT(*impl);
-
-    signable->allocator = allocator;
-    signable->vtable = &s_signable_http_request_vtable;
-    signable->impl = impl;
-
-    /*
-     * Copy the headers since they're not different types
-     */
-    size_t header_count = aws_http_message_get_header_count(request);
-    if (aws_array_list_init_dynamic(
-            &impl->headers, allocator, header_count, sizeof(struct aws_signable_property_list_pair))) {
-        goto on_error;
-    }
-
-    for (size_t i = 0; i < header_count; ++i) {
-        struct aws_http_header header;
-        aws_http_message_get_header(request, &header, i);
-
-        struct aws_signable_property_list_pair property = {.name = header.name, .value = header.value};
-        aws_array_list_push_back(&impl->headers, &property);
-    }
-
-    impl->request = request;
-
-    return signable;
-
-on_error:
-
-    aws_signable_destroy(signable);
-
-    return NULL;
-}
-
-/*
- * This is a simple aws_signable wrapper implementation for an s3 chunk
- */
-struct aws_signable_chunk_impl { /* <------------------ reference here -------------------- */
-    struct aws_input_stream *chunk_data;
-    struct aws_string *previous_signature;
-};
-
-static int s_aws_signable_chunk_get_property(
-    const struct aws_signable *signable,
-    const struct aws_string *name,
-    struct aws_byte_cursor *out_value) {
-
-    struct aws_signable_chunk_impl *impl = signable->impl;
-
-    AWS_ZERO_STRUCT(*out_value);
-
-    /*
-     * uri and method can be queried directly from the wrapper request
-     */
-    if (aws_string_eq(name, g_aws_previous_signature_property_name)) {
-        *out_value = aws_byte_cursor_from_string(impl->previous_signature);
-    } else {
-        return AWS_OP_ERR;
-    }
-
-    return AWS_OP_SUCCESS;
-}
-
-static int s_aws_signable_chunk_get_property_list(
-    const struct aws_signable *signable,
-    const struct aws_string *name,
-    struct aws_array_list **out_list) {
-    (void)signable;
-    (void)name;
-    (void)out_list;
-
-    return AWS_OP_ERR;
-}
-
-static int s_aws_signable_chunk_get_payload_stream(
-    const struct aws_signable *signable,
-    struct aws_input_stream **out_input_stream) {
-
-    struct aws_signable_chunk_impl *impl = signable->impl;
-    *out_input_stream = impl->chunk_data;
-
-    return AWS_OP_SUCCESS;
-}
-
-static void s_aws_signable_chunk_destroy(struct aws_signable *signable) {
-    if (signable == NULL) {
-        return;
-    }
-
-    struct aws_signable_chunk_impl *impl = signable->impl;
-    if (impl == NULL) {
-        return;
-    }
-
-    aws_string_destroy(impl->previous_signature);
-
-    aws_mem_release(signable->allocator, signable);
-}
-
-static struct aws_signable_vtable s_signable_chunk_vtable = {
-    .get_property = s_aws_signable_chunk_get_property,
-    .get_property_list = s_aws_signable_chunk_get_property_list,
-    .get_payload_stream = s_aws_signable_chunk_get_payload_stream,
-    .destroy = s_aws_signable_chunk_destroy,
-};
-
-struct aws_signable *aws_signable_new_chunk(
-    struct aws_allocator *allocator,
-    struct aws_input_stream *chunk_data,
-    struct aws_byte_cursor previous_signature) {
-
-    struct aws_signable *signable = NULL;
-    struct aws_signable_chunk_impl *impl = NULL;
-    aws_mem_acquire_many(
-        allocator, 2, &signable, sizeof(struct aws_signable), &impl, sizeof(struct aws_signable_chunk_impl));
-
-    if (signable == NULL || impl == NULL) {
-        return NULL;
-    }
-
-    AWS_ZERO_STRUCT(*signable);
-    AWS_ZERO_STRUCT(*impl);
-
-    signable->allocator = allocator;
-    signable->vtable = &s_signable_chunk_vtable;
-    signable->impl = impl;
-
-    impl->chunk_data = chunk_data;
-    impl->previous_signature = aws_string_new_from_array(allocator, previous_signature.ptr, previous_signature.len);
-    if (impl->previous_signature == NULL) {
-        goto on_error;
-    }
-
-    return signable;
-
-on_error:
-
-    aws_signable_destroy(signable);
-
-    return NULL;
-}
